Improving injection of rate limiter
